+++ scripts/dist_mobile_build.sh
@@ -11,15 +11,9 @@ cd "$DIR"
 
 export XDG_CACHE_HOME=/tmp/.cache.$$
 
-# Get Go deps
-echo "USER: `id -u $USER`"
-mkdir -p glide_cache
-glide --home "glide_cache" install
-rm -rf glide_cache
-
 # Build!
 echo "==> Building..."
 
-$(which gomobile) bind -v -target=android -tags=mobile -o /workspace/go/src/github.com/mosaicnetworks/babble/build/pkgmobile/mobile.aar github.com/mosaicnetworks/babble/src/mobile 
+$(which gomobile) bind -v -target=android -tags="mobile" -o /workspace/go/src/github.com/mosaicnetworks/babble/build/pkgmobile/mobile.aar github.com/mosaicnetworks/babble/src/mobile 
 
 exit 0